+++ device-minion-runtime/src/shared/minion_rt_helpers/include/common_utils.h
@@ -9,7 +9,7 @@
 * agreement/contract under which the program(s) have been supplied.
 *
 ************************************************************************/
-/*! \file cm_mm_defines.h
+/*! \file common_utils.h
     \brief A C header that defines common utilities that can be used.
 */
 /***********************************************************************/
@@ -20,18 +20,50 @@
 #include <stdio.h>
 #include <inttypes.h>
 
-/*! \enum mprot_field_e
-    \brief enum defining MPROT register fields.
+/*! \def CM_DEFAULT_TRACE_THREAD_MASK
+    \brief Default masks to enable Trace for all hart in CM Shires.
 */
-enum mprot_field {
-    MPROT_FIELD_IO_ACCESS_MODE = 0,
-    MPROT_FIELD_DISABLE_PCIE_ACCESS,
-    MPROT_FIELD_DISABLE_OSBOX_ACCESS,
-    MPROT_FIELD_DRAM_SIZE,
-    MPROT_FIELD_SECURE_MEMORY
-};
+#define CM_DEFAULT_TRACE_THREAD_MASK 0xFFFFFFFFFFFFFFFFULL
 
-typedef uint64_t mprot_field_e;
+/*! \def CM_DEFAULT_TRACE_SHIRE_MASK
+    \brief Default masks to enable Trace for all CM shires.
+*/
+#define CM_DEFAULT_TRACE_SHIRE_MASK 0x1FFFFFFFFULL
+
+/*! \def MM_SHIRE_MASK
+    \brief Master Minion Shire mask
+*/
+#define MM_SHIRE_MASK (1ULL << 32)
+
+/*! \def CM_SHIRE_MASK
+    \brief Shire mask of Compute Minions Shires.
+*/
+#define CM_SHIRE_MASK 0xFFFFFFFFUL
+
+/*! \def CM_MM_SHIRE_MASK
+    \brief Shire mask of Compute Minions and Master Shire.
+*/
+#define CM_MM_SHIRE_MASK 0x1FFFFFFFFULL
+
+/*! \def MM_HART_MASK
+    \brief Master Minion Hart mask
+*/
+#define MM_HART_MASK 0xFFFFFFFFUL
+
+/*! \def CM_HART_MASK
+    \brief Compute Minion Hart mask
+*/
+#define CM_HART_MASK 0xFFFFFFFFFFFFFFFFULL
+
+/*! \def CW_IN_MM_SHIRE
+    \brief Computer worker HART index in MM Shire.
+*/
+#define CW_IN_MM_SHIRE 0xFFFFFFFF00000000ULL
+
+/*! \def MM_HART_COUNT
+    \brief Number of Harts running MMFW.
+*/
+#define MM_HART_COUNT 32U
 
 /*! \fn static inline uint32_t get_set_bit_count(uint64_t mask)
     \brief Count number of set bits in given bit mask
@@ -86,6 +118,22 @@ static inline uint32_t get_msb_set_pos(uint64_t value)
     }
 
     return msb_pos;
+}
+
+/* Function to get the number of enabled harts in given shire and thread mask */
+static inline uint32_t get_enabled_umode_harts(uint64_t shire_mask, uint64_t thread_mask)
+{
+    /* Extract CM shires enabled harts first */
+    uint32_t harts = get_set_bit_count(shire_mask & CM_SHIRE_MASK) * get_set_bit_count(thread_mask);
+
+    /* Check if master shire is enabled */
+    if (shire_mask & MM_SHIRE_MASK)
+    {
+        /* Extract and add upper 32 harts for master shire */
+        harts += get_set_bit_count(thread_mask & CW_IN_MM_SHIRE);
+    }
+
+    return harts;
 }
 
 #endif /* COMMON_UTILS_H */